Conflicts:
	src/sardana/taurus/qt/qtgui/extra_macroexecutor/macrobutton.py
	src/sardana/taurus/qt/qtgui/extra_macroexecutor/ui/macrobutton.ui
	src/sardana/taurus/qt/qtgui/extra_macroexecutor/ui_macrobutton.py
	src/sardana/taurus/qt/qtgui/extra_pool/motor.py
	src/sardana/taurus/qt/qtgui/extra_pool/poolmotor.py
	src/sardana/taurus/qt/qtgui/extra_pool/ui/poolioregisterbuttons.ui
	src/sardana/taurus/qt/qtgui/extra_pool/ui_poolioregisterbuttons.py
	src/sardana/taurus/qt/qtgui/extra_pool/ui_poolmotorslim.py
	src/sardana/taurus/qt/qtgui/extra_pool/ui_taurusmotorh.py
	src/sardana/taurus/qt/qtgui/extra_pool/ui_taurusmotorh2.py
	src/sardana/taurus/qt/qtgui/extra_pool/ui_taurusmotorv.py
	src/sardana/taurus/qt/qtgui/extra_pool/ui_taurusmotorv2.py
	src/sardana/taurus/qt/qtgui/extra_sardana/expdescription.py
	src/sardana/taurus/qt/qtgui/extra_sardana/ui/ui_ExpDescriptionEditor.py
	taurus/lib/taurus/external/unittest/__init__.py
	taurus/lib/taurus/qt/qtgui/extra_guiqwt/curvesmodel.py
	taurus/lib/taurus/qt/qtgui/extra_macroexecutor/ui/MacroButton.ui
	taurus/lib/taurus/qt/qtgui/extra_macroexecutor/ui/macrobutton.ui
	taurus/lib/taurus/qt/qtgui/extra_pool/ui/PoolIORegisterButtons.ui
	taurus/lib/taurus/qt/qtgui/extra_pool/ui/poolioregisterbuttons.ui
	taurus/lib/taurus/qt/qtgui/panel/qdataexportdialog.py
	taurus/lib/taurus/qt/qtgui/panel/qdoublelist.py
	taurus/lib/taurus/qt/qtgui/panel/report/basicreport.py
	taurus/lib/taurus/qt/qtgui/panel/taurusconfigurationpanel.py
	taurus/lib/taurus/qt/qtgui/panel/taurusinputpanel.py
	taurus/lib/taurus/qt/qtgui/panel/test/__init__.py
	taurus/lib/taurus/qt/qtgui/plot/arrayedit.py
	taurus/lib/taurus/qt/qtgui/plot/curveStatsDlg.py
	taurus/lib/taurus/qt/qtgui/plot/curvesAppearanceChooserDlg.py
	taurus/lib/taurus/qt/qtgui/plot/qwtdialog.py
	taurus/lib/taurus/qt/qtgui/plot/taurusplotconf.py
	taurus/lib/taurus/qt/qtgui/taurusgui/taurusgui.py

+++ src/sardana/pool/poolcontrollers/DiscretePseudoMotorController.py
@@ -77,7 +77,7 @@ class DiscretePseudoMotorController(PseudoMotorController):
     def GetAxisAttributes(self, axis):
         axis_attrs = PseudoMotorController.GetAxisAttributes(self, axis)
         axis_attrs = dict(axis_attrs)
-        axis_attrs['Position']['type'] = int
+        axis_attrs['Position']['type'] = float
         return axis_attrs
